+++ pages/contact.js
@@ -1,9 +1,29 @@
+import { useState } from 'react';
+import { ContactUsFormSubscribe } from '@/components/ContactUsForm';
+import contactUsFormStyles from '@/styles/Contact.module.scss';
+import Bracket from '@/components/decorations/Bracket';
+
 export default function ContactUs() {
+  const [message, setMessage] = useState([]);
+
   return (
-    <div className='contact-us'>
-      <h2>Contact Us</h2>
-      <p>We're building this page.</p>
-      <p>Stay tuned!</p>
+    <div className={contactUsFormStyles.contact}>
+      <div className={contactUsFormStyles.contact__formAndDecorations}>
+        <Bracket className={contactUsFormStyles.contact__yellowBracket} />
+        <ContactUsFormSubscribe setMsg={setMessage} />
+        <img
+          src='/images/svg/yellow-colon.svg'
+          className={contactUsFormStyles.contact__yellowColon}
+        />
+        <div className={contactUsFormStyles.contact__response_message}>
+          {message?.map((m, i) => (
+            <span key={i}>
+              {m}
+              <br />
+            </span>
+          ))}
+        </div>
+      </div>
     </div>
   );
 }